requirements.txt
# PHOENIX Sign Language Recognition Requirements

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Utilities
tqdm>=4.65.0
einops>=0.6.0
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web interface (optional)
streamlit>=1.25.0
plotly>=5.15.0

# Speech-to-Sign (optional)
# pip install openai-whisper  # For speech input
# pip install transformers    # Alternative for German text processing

